(DOCS) Add example app
This commit adds an example implementation of a fictional TSToy application,
mirroring the fictional example used in the DSCv2 docs. The implementation
is used for introspection and to serve as a functional target for the example
DSC Resource implementations.

The fictional application has two configuration files that can be defined to
control its behavior, one in the machine scope and another in the user scope.

Users can set the value of `updates.automatic` to `true` or `false` and the
value of `updates.checkFrequency` to an integer between `0` and `90`, inclusive.

The implementation has the `tstoy show` command, which returns the current
configuration information for the application. It shows the values for:

- The default settings, before any settings from the configuration files or user
  input are processed.
- The settings defined in the machine-scope configuration file.
- The settings defined in the user-scope configuration file.
- The final settings, after merging the defaults, configuration files, and any
  user-input, including environmental variables.

You can specify a smaller list of outputs with the `--only` flag. You can also
run the `tstoy show path` command to enumerate the paths to where it expects to
find configuration files. By default, it shows the machine-scope file path and
then the user-scope file path. If you specify either `machine` or `user` as an
argument to the command, it returns only that path.

For example:

```sh
tstoy show path machine
```

```output
C:\ProgramData\TailSpinToys\tstoy\tstoy.config.json
```

A future commit will implement a DSC Resource against the fictional application.