(DOCS) Add initial implementation of go resource
This commit adds the initial implementation of a DSC Resource for managing
the fictional TSToy app in golang. It is only implemented for the get
method in this initial implementation.

You can pass JSON blobs to the command from stdin, which it will use instead
of the flags `--scope`, `--ensure`, `--updateAutomatically`, and
`--updateFrequency`.

The get command is implemented to return the current configuration settings
for a single file unless you use the `--all` flag, in which case it returns
the configuration settings for both scopes.

The get command returns the configuration settings as JSON blobs, one per
line.

More work is required to implement the test and set commands and to author
the resource manifest.