(DOCS) Implement set command and reference docs
This change implements the `set` command for the go implementation and adds
reference documentation and convenience UX functionality to the application.

It also defines an initial resource manifest and updates the build to include
the manifest in the output archive and in the dist path when building only
for the current OS/architecture.

This change also adds a build script to the example folder itself so the app
and implementations can be built together and then used. It includes a brief
maintainer-facing README so folks can investigate the implementation.

It does not yet include narrative documentation or proper tests. More work
is required for verifying its use with DSC.