(DOCS) Update readme with `dsc` calls
This change updates the examples README to show how you can use the
`gotstoy` app as a DSC Resource. It also temporarily updates the
resource manifest to set `preTest` to `true` so DSC does not throw
an error about the test method not being implemented.

It shows using `gotstoy` with `dsc config get`. The other commands
aren't implemented yet.