(DOCS) Add _Write a DSC Resource in Go_ tutorial
This change builds on the example implementations to present a tutorial
for writing a command-based DSC Resource in Go. It includes an overview
of the fictional TSToy application, which the tutorial's DSC Resource
manages.

While this first tutorial is written in Go, it should be one of several
tutorials that show how developers can implement command-based DSC
Resources in their language of choice.

Eventually, the fictional application and sample code for these
tutorials should move out into a separate repository. For now, keeping
it in one place is convenient.