(DOCS) (WIP) Define & doc resource manifest schema
This commit converts the type definitions in the
`dsc_lib\src\dscresources\resource_manifest.rs` file into a minimal
JSON schema. The schema represents the current state of the type
definitions, but raises a few questions for discussion.

The documentation for the schema is currently hand-authored and the
structure is experimental, since we have no official support for
documenting JSON schemas on the Learn platform.

For readability and referability, I recommend breaking the larger
schema up into smaller schemas and referencing across them.