(DOCS) Define decomposed schemas
This change defines a set of JSON Schema documents for DSC. Instead of a
few monolithic schema documents, this change defines the schemas as
smaller, more manageable documents that can be bundled as needed.

These schemas are hand-authored and more specific than the schemas
generated with the `dsc schema` command. The schemas included here
are implemented with loose validation, but we can support strict
validation side-by-side.

The schemas are organized into subfolders:

- In the `bundled` folder, the schema documents include a synthetic
  example of the bundling process. For this change, they were generated
  by hand, but they could be composed with tooling instead.
- In the `config` folder, the schema documents validate a DSC Configuration
  document.
- In the `definitions` folder, the schema documents represent reused
  definitions for validation and/or documentation. The `title` and
  `description` keywords for instances in a JSON Schema are used by
  authoring tools to provide contextual documentation.
- In the `outputs` folder, the schema documents represent the output
  users and integrating tools should expect from calling `dsc`. Tools
  may use these definitions to generate their own structs in their
  application code to more readily and accurately handle output from
  DSC.
- In the `resource` folder, the `manifest*` schema documents validate
  a command-based DSC Resource manifest.
- In the `resource/properties` folder, the schema documents represent
  standard DSC Resource properties that have well-defined usage.

The schema URL and the ID names for the schemas is best-effort but
should not be treated as canonical without further discussion.

The documents include comments indicating ambiguities or questions
that they raise.