(MAINT) Add build script for bundling schemas
This change adds a new build script to the schema folder. By default,
it generated bundled schema files for the configuration document and
resource manifest schemas. It can bundle schemas in four formats:

- JSON
- YAML
- JSON (VS Code)
- YAML (VS Code)

The VS Code bundles have their definitions organized differently -
VS Code has not yet fully implemented 2020-12 support, so it can not
resolve non-local (ie not starting with `#/`) references correctly.

The non-VS Code bundles were validated with [HyperJump][01].

[01]: https://json-schema.hyperjump.io/

This change also temporarily includes the gotstoy schema in the bundled
output folder so you can review how the schema affects the authoring
DevX.

This change includes a bundle config file for determining which schemas
should be bundled and how they should be named. While the outputs are
committed here for convenience and reviewability, they should instead
be generated with the build script and published as needed. They're
build artifacts generated from the decomposed schema definitions.

The build step doesn't support injecting links to the documentation,
but we should add that functionality in the future. This would make it
easier for authors and editors of configuration documents and resource
manifests to look up the documentation without having to search for it.