(AB-118300) Draft Anatomy of a command-based DSC Resource
This change adds a draft of the conceptual article defining the components of
a command-based DSC Resource at a high-level, with information about what is
required of them and how DSC uses them.

This draft links to several articles that have not been written yet. Those
drafts will be addressed in future PRs.