(MAINT) Simplify decomposed schemas
This change rearranges the definitions for configuration parameters and
resources so that only the valid parameter/resource definition in the
`parameters` and `resources` properties are references. This makes the
documentation and definition a little cleaner and easier to maintain.

The validation properties are moved from definitions into the parameter
document for the same reason - they aren not referenced anywhere else.

In the resource instance schema, this change adds a validation pattern
for the `dependsOn` key to match the proposed syntax.

In the resource manifest schema, this change ensures that tags for a
resource must be unique, disallowing repeated tags.