(GH-96) Replace jsonschema crate with boon
Prior to this change, the DSC library, CLI, and the
Registry resource all relied on the `jsonschema` crate
for schema compilation and validation.

As raised in #96, the `jsonschema` crate does not
support newer versions of JSON Schema like drafts
2019-09 and 2020-12. Further, the implementation
fails several of the tests designed by the JSON
Schema organization to validate implementations
for the newest dialect  it _does_ support (Draft 7).

This change prototypes the replacement of the
`jsonschema` crate with the `boon` crate, which
has full support for every released draft of
JSON Schema and currently passes every test
for implementation validation.

In this initial prototype change, the code
includes duplicated behaviors that can be
cleaned up or abstracted into helper
functions.

Another benefit and possible avenue going
forward is to turn the decomposed schemas
in the repository into the source of truth
for validating the incoming data. With
Boon, DSC could pre-load the schemas and
only have to compile schemas for resources
on-demand thereafter.

In that model, the JSON schemas in the
repository could be embedded in the binary
or shipped alongside it.

This change also updates the use of
Schemars to output Draft 2019-09 schemas
for DSC and the Registry resource, to
make them closer to the definitions for
the schemas in the repository.

This commit adds documentation to the
structs in the Registry resource, showing
how they can be used to provide improved
documentation for the schema at author-time.