(GH-178) Document `dsc` exit codes
Prior to this change, the documentation for the `dsc` command did not
include the exit codes that it may return. This change adds a section
to the documentation that lists the exit codes and their meanings.

This information is particularly important to integrating tools, like
winget, that need to know the semantic meaning of the exit codes when
calling `dsc`.

- Resolves #178