(MAINT) Update documentation and schema for changes from #175
Prior to this change, the schema for the `dependsOn` keyword used the PSDSC
syntax and the schema for instance names allowed arbitrary strings.

This change:

- Updates the schema and documentation for instance names to clarify that
  they must be non-empty strings containing only letters, numbers, and
  spaces.
- Updates the schema and documentation for the `dependsOn` property to
  use the new `resourceId()` function syntax as added in #175.

For now, this change does not define new reference documentation for the
configuration functions. When more configuration functions are added, we
can define a separate set of reference documentation for those functions.