(GH-189) Add semantic changelog
Prior to this change, the project didn't maintain a changelog file. Instead, it
relied on generation of release notes from PRs.

As raised in #189, the auto-generated changelog is limited in its effectiveness
for changelog readers outside of the maintenance team. The entries depend
entirely on the PR titles, they aren't organized by kind or relative importance,
and they can't link to documentation.

This change:

- Defines the changelog by hand, adhering to the KAC format with an extension for
  how related work items are displayed, using a details element to collapse them.
- Fully documents the set of user-facing changes that are included in the `alpha.2`
  release, including summaries and related work items and linking to documentation
  where appropriate.
- Resolves #189