(MAINT) Remove markdown description from semver
Prior to this change, the markdown description for the semver schema was
used anywhere the value of a property referenced that definition instead
of the defined description on that property.

This change converts the markdownDescription for the definition into a
comment for maintainers, ensuring that more specific documentation can
be shown in the hover context for manifest and configuration authors.