(MAINT) Fix schema gen for array values
Prior to this change, the schema generation had a bug that incorrectly
munged the values for always-array keywords, like `enum` and `requires`
to scalar values when the definition was a single-item array.

This change ensures that those values are not incorrectly munged, fixing
hidden bugs in the behavior for those values in the generated schemas.