(SCHEMA) Replace `args` inputKind with `env`
This change updates the schemas to replace the unimplemented `args` option
for command input in the resource manifest with the `env` option. It also
updates the schema and documentation to indicate that when the `input` option
is not defined for a command in a resource manifest, DSC does not send the
resource any input for that command.

This update is made to reflect the changes from #198.