(MAINT) Clarify `env` input handling
Prior to this change, the enum documentation for the enhanced schema
didn't clarify how the environment variables are named or their values
passed.

This change clarifies both, making it easier for manifest authors to
determine whether the input option is right for them and how they should
expect to handle the input in their resource without consulting the
online documentation.