(MAINT) Fix schema docs for `input` in set/test
Prior to this change, the schema docs erroneously noted that
when `input` is not defined in a resource manifest for the
`set` and `test` definitions that DSC doesn't send any input.

This is _technically_ true, but the `input` property is included
in the list of `required` properties for `set` and `test`, unlike
`get`.

This change removes the misleading sentence, since those properties
can't be undefined in a valid manifest.