(GH-224) Convert schema constants to enums
This change converts the string constants for the canonical schema URIs
for both the configuration document and resource manifest schemas to
enums, defining the valid URIs as variants.

For each supported version of a schema, there are three variants:

1. `Version<YYYY>_<MM>` - The canonical URI to the non-bundled schema.
    Using this schema requires retrieving every referenced schema, but
    it is also the least-munged.
1. `Bundled<YYYY>_<MM>` - The canonical URI to the bundled schema. When
   using this schema, only the bundled schema needs to be retrieved.
1. `VSCode<YYYY>_<MM>` - The canonical URI to the enhanced authoring
   schema. This schema is specifically implemented to support contextual
   help and DevX when authoring in VS Code. It is much larger than the
   other schemas.

Which gives six variants:

- `Version2023_08`, `Bundled2023_08`, and `VSCode2023_08` for 2023/08
- `Version2023_10`, `Bundled2023_10`, and `VSCode2023_10` for 2023/10

Currently, the specified schema is not actually used by DSC to validate
the document or manifest. In the future, we could use the version to
identify how to validate and process the data.