(MAINT) Update GitHub templates
This change copies the default templates for PowerShell repositories
and customizes them for DSC.

In particular, the bug report template is updated to more usefully
reflect errors and issues with DSC, rather than a PowerShell command.

The change includes the feature request template and template config,
because adding any files to the `.github/ISSUE_TEMPLATE` directory
requires all of them to be present - none of the defaults in the
community repository are used when this folder exists.

This change also updates the workflow file to avoid triggering rust
build and tests when the PR only edits the issue templates.