(SCHEMA) Update schemas for the provider->adapter rename
This change updates the source schemas to account for renaming DSC Resource
Providers to DSC Resource Adapters in #334. This is a breaking schema change
that requires a new canonical URI, so the change also adds new schema URIs to
DSC to account for the changes.

The regenerated schemas will be added in the next commit.