(SCHEMA) Update schemas for `metadata` in output
This changes defines the metadata properties that DSC returns in the
results for the `dsc config *` commands, as implemented in #405.

It defines each metadata property as a separate file in a new folder
for metadata schemas for easier reuse, as well as two definitions
of the `metadata` property used across different result objects, one
for the top-level of a configuration and one for resource instances
in that result.

This change also required a minor fix to the build script to account
for array items that are objects solely defined by a `$ref` keyword
and to find and compose schemas from the new `metadata` folder.

The change includes the rename for the `requiredSecurityContext`
metadata property to `securityContext`.