(SCHEMA) Add VSCode schemas for output and update examples
Prior to this change, the output schemas were bundled, but not
with in the format compatible with VS Code.

This change adds the VS Code compatible bundle format to the
composed schemas configuration. This enables us to show new
examples of output data with IntelliSense to make inspecting
the output structure easier and more interactive.

This change also reorganizes the schema examples folder to
use subfolders for configuration documents, resource manifests,
and the new output data objects.