(GH-642) Generate schemas from source for v3.0
This change generates the canonical schemas for in new folders: `v3`,
`v3.0`, and `v3.0.0`, instead of the previous date-based folder
`2024/04`. The changing of the casing is _not_ backwards compatible, and
requires a new location to avoid breaking prior schemas.

This change generates the schemas into three folders to accomodate the
changes requested in #138 for moving from date-based to semantic
versioning for the schemas.

Going forward, schema generation should:

- Export to a folder for the current release with the full semantic
  version, like `v3.0.1` or `v3.1.0`.
- Export to a folder for the current `major.minor` schema, like `v3.0` or
  `v3.1`
- Build to a folder for the current major version of the release, like `v3`.

Under this design, you can:

- Get the exact version of the schema that a release shipped with by
  specifying the schema URI as `v<major>.<minor>.<patch>`
- Get the latest schema for your minor version as `v<major>.<minor>`
- Get the latest schema for your major version as `v<major>`

Breaking changes to the schema have the potential to break
integrating software, configuration documents, and resource manifests.
They therefore constitute a breaking change in the application and
require incrementing the major version segment of the semantic version.

This change will be followed by an update to the allowed schema versions
in the code itself, which should only recognize the updated schema at
this time, not previous schemas.