(GH-642) Add schemas module, update for schema handling in v3
This commit only touches the `dsc` and `dsc_lib` crates. A future
commit will update the examples, manifests, and tests throughout
the repository for the updated schema URIs. Tests against this
commit specifically will not pass, but the changes to the library
are broken out here for easier review.

This change defines a new module in the library for interacting
with and managing JSON Schemas for DSC. The code in this module
is made public for use in other repository crates, but shouldn't
be used by crates outside of this repository.

It defines new enumerations:

- `SchemaUriPrefix` for handling URIs pointing to either the GitHub repository
  or friendly `aka.ms` locations.
- `SchemaForm` for the different forms the DSC schemas are published to (canonical,
  bundled, and enhanced for VS Code)
- `RecognizedSchemaVersion` for managing the version folders schemas
  are publsihed to for and after GA.

It defines a few helper functions for constructing the various URIs for a
schema. These helper functions may be useful for testing, but should
primarily be used through the new `DscRepoSchema` trait, which requires
a type to define the base name and folder path for its schema and whether
the schema should be published in its bundled form (only applies to top-level
schemas, not subschemas).

The trait provides default implementations for retrieving the schema URIs and
a function to return a generated JSON Schema for the `$schema` keyword, used
by both configuration documents and resource manifests.

As we canonicalize and improve the schemas, this trait should make it simpler
to manage the schemas across new versions and should be the container for
JSON Schema enhancements specific to the DSC schemas.

This change updates the configuration document and resource manifest definitions
to implement the `DscRepoSchema` trait and replaces the per-type enumerations
for schema version URIs by passing the `recognized_schema_uris_subschema` to
the schemars `schema_with` attribute.