(GH-642) Update schema URIs to aka.ms and v3
This change updates the schema URIs in the various resource manifests,
configuration documents, and tests to use the new schema URIs:

- `https://aka.ms/dsc/schemas/v3/bundled/config/document.json` for
  configuration documents.
- `https://aka.ms/dsc/schemas/v3/bundled/resource/manifest.json` for
  resource manifests.

This change pins the changes to the major version to reduce how often
these files need to be updated with future releases. It uses the
`aka.ms` prefix instead of the GitHub URI prefix for readability
and because we expect the majority of users to prefer the shorter
URI.

With this change in place, tests broken in the previous commit should
pass again.