(GH-642) Add initial plumbing for validating schema URIs
Prior to this change, DSC didn't have a convenient way to verify
whether a configuration or manifest used a recognized schema or
to switch behaviors based on the declared URI.

This change begins the process of plumbing the recognized schemas
into DSC so the engine can correctly validate and handle items
based on the value of the `$schema` keyword. In this first
iteration, the `DscRepoSchema` trait adds a method to validate
whether the item uses a recognized schema. In the future, the
engine should provide further handling than pass/fail.

This change only adds the methods, it doesn't alter how DSC
processes configuration documents or resource manifests.