(GH-642) Document schema URI publishing and update manifest/config ref
This change drafts a new reference document that explains how the DSC
schemas are published, the URIs DSC recognizes for those schemas, and
updates the documentation for the resource manifest and configuration
document schemas with the updated context.

A future change will address updating the rest of the reference
documentation.