(SCHEMA) Decompose capabilities into separate schema for reuse
This change breaks out the capabilities schema from the `dsc resource list`
output schema into a separate definition file. This enables us to reference
the schema without using a JSON Pointer and provides us with a page for
documenting the capabilities directly.

This is a required precursor for a future change that defines stdout
schemas for resource operations.