(SCHEMA) Add stdout schemas for resource operation commands
Prior to this change, the instructions and information about how
resources should return data to DSC for any given operation was
somewhat vague and not always helpful.

This change defines a new set of schemas for the various operations,
so resource authors can review their implementations and validate them
against these schemas.