(DOCS) Scaffold resource authoring docs
This change adds empty documents to suggest the structure
and content of a resource authoring guide.

It is intended to convey high-level design guidance and
implementation-agnostic guidance for authoring resource
manifests and instance schemas.

It doesn't cover any specifics for defining commands or
the resource in a particular language. Those concerns
are better covered in the resource development kits or
the sample resource tutorials, not the conceptual guide.