(DOCS) Draft resource design docs
This change adds conceptual documentation as the initial offering for
a resource _design_ guide.

It is intended to convey high-level  implementation-agnostic guidance
for defining DSC resources, including resource manifests and instance
schemas.

The guide doesn't cover any specifics for defining commands or the
resource in a particular language. Those concerns are better covered in
the resource development kits or the sample resource tutorials, not the
conceptual guide.