(GH-1153) Move library crates into `lib` folder
This change moves the library crates, which don't have an executable,
into the `lib` folder. This makes discovering shared libraries easier,
instead of relying on the `_lib` convention the crates use. This change
doesn't rename the libraries, it only moves them.

This change doesn't require modifying any code or cargo manifests,
except for pointing to the correct path in the virtual workspace.