(GH-1153) Move adapter subprojects into `adapters` folder
This change moves the adapter subprojects, which have an executable,
into the `adapters` folder. This makes discovering which projects
provide DSC resource adapters easier, instead of relying on the
`-adapter` suffix convention. This change does rename the folders,
dropping the suffix. The adapters aren't rust crates, so this doesn't
affect code compilation.

This change doesn't require modifying any code or cargo manifests,
except for pointing to the correct path in the build script.