(GH-1153) Move tree-sitter subprojects into `grammars` folder
This change moves the tree-sitter subprojects, which define grammars,
into the `grammars` folder. This makes discovering which projects
provide DSC grammars easier, instead of relying on the `tree-sitter-`
prefix convention. This change does not rename the folders, it only
moves them.

This change doesn't require modifying any code or cargo manifests,
except for pointing to the correct path in the build script and
virtual workspace manifest.

This change also updates the `.gitignore` to generically ignore
generated grammar files in the `grammars` folder.