(GH-1153) rename `dsc_lib` to `dsc-lib`
This change renames `dsc_lib` to `dsc-lib`. This is done for
a few reasons:

1. The general naming convention for rust crates is `kebab-case` for
   crates, `snake_case` for modules. This aligns the naming convention.

   Note that kebab-case crate names are still referenced with underscores
   as `snake_case`, so no code needed to be changed in crates, only the
   cargo manifests.
1. When we eventually publish the library crates, we will need to use a
   coherent naming convention. Standardizing on `dsc-lib` as a prefix
   ensures that the other library crates can follow this convention and
   more readily be discovered and collected. The other libraries are only
   intended for use with DSC.

This change ensures that the manifests and build script are updated to use
the renamed crate. I decided to keep a 1:1 map of crate name and folder,
to limit confusion over having more than one `dsc` folder in the project.