(GH-1153) Rename `pal` to `dsc-lib-pal`
This change renames the `pal` crate to `dsc-lib-pal`. This is to
support eventual publishing of the crates. When we implement the
Rust RDK, we will need to publish the library crates, and this
ensures the naming is both consistent and easily discoverable.

This change makes the requisite updates to cargo manifests, Rust
code, and the build script.