(GH-1153) Rename `security_context_lib` to `dsc-lib-security_context`
This change renames the `security_context_lib` crate to
`dsc-lib-security_context`. This is to support eventual
publishing of the crates. This crate is a dependency not
only for the `registry` resource but also the `dsc-lib`
crate. When we implement the Rust RDK, we will need to
publish the library crates, and this ensures the naming
is both consistent and easily discoverable.

This change makes the requisite updates to cargo manifests, Rust
code, and the build script.