(GH-1152, GH-1153) Add new build script and workflow
This change adds a reworked build script to the project and
a GHA workflow that uses it. Currently this duplicates both
the build script and the CI jobs. For now, this enables us
to ensure continuity of existing behavior while we iteratively
improve the new build script and CI configuration to speed up
builds and improve maintainability.

The new build script generally provides equivalent functionality
with the existing legacy script to the best of my ability.

In the process, this change also:

- Defines a helper module for build-and-maintenance related functions.
  This made the build script itself somewhat easier to reason about,
  and the functions in the module can be used outside of the script.
- Moves build data out of the script and into a separate datafile. For
  now, this puts build data at the root, instead of per-project. In the
  future, I would like to refactor things to retrieve build data from
  the individual projects instead.

Finally, this change adds a benchmarking comparison script to enable us
to programmatically compare the build scripts for local and CI workflows.